pyproject.toml
[tool.poetry]
name = "fin-qdrant-rag"
version = "0.1.0"
description = "RAG system for finance/trading PDFs with FastAPI, Qdrant, OpenAI, and LangChain."
authors = ["Lucas <lucaimbalzano@gmail.com>"]
readme = "README.md"
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
qdrant-client = "^1.7.0"
openai = "^1.14.3"
langchain = "^0.3.26"
pymupdf = "^1.23.22"
pdfplumber = "^0.10.3"
apscheduler = "^3.10.4"
python-dotenv = "^1.0.1"
pydantic = "^2.6.4"
asyncpg = "^0.29.0"
greenlet = "^3.0.3"
redis = "^5.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
httpx = "^0.27.0"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"